requirements.txt
Pillow==3.0.0
PyAudio==0.2.9
wheel==0.24.0
youtube-dl==2015.11.27.1
google-api-python-client==1.4.2
